pyproject.toml
[project]
name = "my-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dotenv>=0.9.9",
    "git-filter-repo>=2.47.0",
    "google-cloud-bigquery>=3.30.0",
    "pandas>=2.0.3",
    "pytest>=8.3.5",
    "requests>=2.32.4",
]

[project.scripts]
my-test = "my_test:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
